requirement.txt
amqp==5.0.6
appdirs==1.4.4
asgiref==3.3.4
astroid==2.5.7
billiard==3.6.4.0
cairocffi==1.2.0
CairoSVG==2.5.2
celery==5.1.1
cffi==1.14.5
click==7.1.2
click-didyoumean==0.0.3
click-plugins==1.1.1
click-repl==0.2.0
cssselect2==0.4.1
defusedxml==0.7.1
Django==3.2.4
django-crispy-forms==1.12.0
django-debug-toolbar==3.2.1
django-filter==2.4.0
django-jalali==4.1.2
django-rest-framework==0.1.0
django-widget-tweaks==1.4.8
djangorestframework==3.12.4
dotenv-python==0.0.1
gitdb==4.0.7
GitPython==3.1.18
gunicorn==20.1.0
html5lib==1.1
isort==5.8.0
itsdangerous==2.0.1
jdatetime==3.6.2
Jinja2==3.0.1
kombu==5.1.0
lazy-object-proxy==1.6.0
MarkupSafe==2.0.1
mccabe==0.6.1
pathspec==0.8.1
pbr==5.6.0
Pillow==8.2.0
prompt-toolkit==3.0.19
psycopg2-binary==2.8.6
py==1.10.0
pycodestyle==2.7.0
pycparser==2.20
pymemcache==3.4.4
pyphen==0.11.0
python-dotenv==0.19.0
pytz==2021.1
regex==2021.4.4
six==1.16.0
smmap==4.0.0
sqlparse==0.4.1
stevedore==3.3.0
tinycss2==1.1.0
toml==0.10.2
vine==5.0.0
wcwidth==0.2.5
WeasyPrint==52.5
webencodings==0.5.1
Werkzeug==2.0.1
wrapt==1.12.1
djangorestframework-simplejwt==4.7.2